refactored column
